# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pulsar_data_aggregation_API",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/data-aggregation-contract-and-fact/get-by-id/{id}": {
            "get": {
                "tags": [
                    "Contract and fact"
                ],
                "operationId": "de9200d0291f7e8155adc4de29e8908d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/data-aggregation-contract-and-fact/get-by-name/{company_name}": {
            "get": {
                "tags": [
                    "Contract and fact"
                ],
                "operationId": "45f6b494ab35b482f45a340f4595cc0d",
                "parameters": [
                    {
                        "name": "company_name",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}